server/requirements.txt
flask
flask-cors
numpy
pandas
numpy
psycopg2-binary~=2.8.6
tqdm